Some cleanup and refactoring, replace custom adapter with generic one
