--- app/src/main/java/eu/kanade/tachiyomi/ui/setting/database/DatabaseSourceAdapter.kt
@@ -1,12 +1,0 @@
-package eu.kanade.tachiyomi.ui.setting.database
-
-import com.bluelinelabs.conductor.Controller
-import eu.davidea.flexibleadapter.FlexibleAdapter
-
-class DatabaseSourceAdapter(controller: Controller) :
-    FlexibleAdapter<DatabaseSourceItem>(null, controller, true) {
-
-    init {
-        setDisplayHeadersAtStartUp(true)
-    }
-}
